requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
Pillow==10.0.1
click==8.1.7
colorama==0.4.6
tqdm==4.66.1
python-dotenv==1.0.0
flask==2.3.3
werkzeug==2.3.7
gunicorn==21.2.0